+++ integrations/shutdown/v1/definition/consts.go
@@ -18,7 +18,7 @@
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
 //
 
-package v1
+package definition
 
 const (
 	Name = "shutdown.v1"