+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - (Feature) Gateway Group for ArangoDeployment
 - (Feature) Gateway config loader
 - (Feature) ConfigV1 Integration Service
+- (Feature) Integration Service Authentication
 
 ## [1.2.42](https://github.com/arangodb/kube-arangodb/tree/1.2.42) (2024-07-23)
 - (Maintenance) Go 1.22.4 & Kubernetes 1.29.6 libraries